requirements.txt
pillow>=10.0
mss>=9.0
pynput>=1.7
pywin32>=300
